- Fixed a couple of uninitialized variables (patch #2239). - Fixed a warning.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@4624 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/src/main.h
@@ -302,7 +302,7 @@ class MainFrame : public wxFrame
         bool m_SmallToolBar;
         bool m_StartupDone;
         bool m_InitiatedShutdown;
-        
+
         bool m_AutoHideLogs;
         int m_AutoHideLockCounter;
 
@@ -313,8 +313,8 @@ class MainFrame : public wxFrame
 
         typedef std::map<int, const wxString> MenuIDToScript; // script menuitem ID -> script function name
 		MenuIDToScript m_MenuIDToScript;
-		
-		typedef std::map<Logger*, size_t> LoggerToInfoPaneIndex; // logger to infpane page index
+
+		typedef std::map<Logger*, int> LoggerToInfoPaneIndex; // logger to infpane page index
 		LoggerToInfoPaneIndex m_LoggerToInfoPaneIndex;
 
         wxDialog* m_pBatchBuildDialog;
